Add abstract types
